Fixed illegal package name
