+++ app/build.gradle
@@ -3,7 +3,7 @@
  */
 
 plugins {
-    id 'java.dj.djm.java-application-conventions'
+    id 'de.dj.djm.java-application-conventions'
 }
 
 dependencies {
@@ -11,13 +11,13 @@ dependencies {
 }
 
 application {
-    mainClassName = 'java.dj.djm.app.App'
+    mainClassName = 'de.dj.djm.app.App'
 }
 
 jar {
     manifest {
         attributes(
-                'Main-Class': 'java.dj.djm.app.App'
+                'Main-Class': 'de.dj.djm.app.App'
         )
     }
 }
