+++ app/src/main/java/de/dj/djm/app/App.java
@@ -1,0 +1,12 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package de.dj.djm.app;
+
+import de.dj.djm.parser.Parser;
+
+public class App {
+    public static void main(String[] args) {
+        System.out.println(Parser.parse("World"));
+    }
+}
