+++ app/src/main/java/de/dj/djm/app/MessageUtils.java
@@ -1,0 +1,10 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package de.dj.djm.app;
+
+class MessageUtils {
+    public static String getMessage() {
+        return "Hello      World!";
+    }
+}
