--- app/src/main/java/java/dj/djm/app/App.java
@@ -1,12 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package java.dj.djm.app;
-
-import java.dj.djm.parser.Parser;
-
-public class App {
-    public static void main(String[] args) {
-        System.out.println(Parser.parse("World"));
-    }
-}
