--- app/src/main/java/java/dj/djm/app/MessageUtils.java
@@ -1,10 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package java.dj.djm.app;
-
-class MessageUtils {
-    public static String getMessage() {
-        return "Hello      World!";
-    }
-}
