+++ app/src/test/java/de/dj/djm/app/MessageUtilsTest.java
@@ -1,0 +1,14 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package de.dj.djm.app;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+class MessageUtilsTest {
+    @Test void testGetMessage() {
+        assertEquals("Hello      World!", MessageUtils.getMessage());
+    }
+}
