--- app/src/test/java/java/dj/djm/app/MessageUtilsTest.java
@@ -1,14 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package java.dj.djm.app;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-class MessageUtilsTest {
-    @Test void testGetMessage() {
-        assertEquals("Hello      World!", MessageUtils.getMessage());
-    }
-}
