+++ base/build.gradle
@@ -3,5 +3,5 @@
  */
 
 plugins {
-    id 'java.dj.djm.java-library-conventions'
+    id 'de.dj.djm.java-library-conventions'
 }
