+++ base/src/main/java/de/dj/djm/base/Base.java
@@ -1,0 +1,7 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package de.dj.djm.base;
+
+public class Base {
+}
