--- base/src/main/java/java/dj/djm/base/Base.java
@@ -1,7 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package java.dj.djm.base;
-
-public class Base {
-}
