+++ base/src/test/java/de/dj/djm/base/BaseTest.java
@@ -1,0 +1,27 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package de.dj.djm.base;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class BaseTest {
+
+    @BeforeEach
+    void setup() {
+
+    }
+
+    @AfterEach
+    void teardown() {
+
+    }
+
+    @Test
+    void test() {
+
+
+    }
+}
