--- base/src/test/java/java/dj/djm/base/BaseTest.java
@@ -1,27 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package java.dj.djm.base;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class BaseTest {
-
-    @BeforeEach
-    void setup() {
-
-    }
-
-    @AfterEach
-    void teardown() {
-
-    }
-
-    @Test
-    void test() {
-
-
-    }
-}
