+++ buildSrc/src/main/groovy/de.dj.djm.java-common-conventions.gradle
