--- buildSrc/src/main/groovy/java.dj.djm.java-application-conventions.gradle
@@ -1,11 +1,0 @@
-/*
- * This file was generated by the Gradle 'init' task.
- */
-
-plugins {
-    // Apply the common convention plugin for shared build configuration between library and application projects.
-    id 'java.dj.djm.java-common-conventions'
-
-    // Apply the application plugin to add support for building a CLI application in Java.
-    id 'application'
-}
