+++ parser/src/main/java/de/dj/djm/parser/Parser.java
@@ -1,0 +1,12 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package de.dj.djm.parser;
+
+public class Parser {
+    public static final String BEGIN_TAG = "<";
+    public static final String END_TAG = "/>";
+    public static String parse(String input) {
+        return "Hello, " + input;
+    }
+}
