+++ parser/src/main/java/de/dj/djm/parser/data/Examples.java
@@ -1,0 +1,15 @@
+package de.dj.djm.parser.data;
+
+public class Examples {
+    public static final String XML = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n" +
+            "<databaseConfig>\n" +
+            "    <database>\n" +
+            "        <name>myDatabase</name>\n" +
+            "        <type>MySQL</type>\n" +
+            "        <host>localhost</host>\n" +
+            "        <port>3306</port>\n" +
+            "        <username>admin</username>\n" +
+            "        <password>password123</password>\n" +
+            "    </database>\n" +
+            "</databaseConfig>";
+}
