--- parser/src/main/java/java/dj/djm/parser/Parser.java
@@ -1,10 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package java.dj.djm.parser;
-
-public class Parser {
-    public static String parse(String input) {
-        return "Hello, " + input;
-    }
-}
