+++ parser/src/test/java/de/dj/djm/parser/ParserTest.java
@@ -1,0 +1,30 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package de.dj.djm.parser;
+
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class ParserTest {
+
+    @BeforeEach
+    void setup() {
+
+    }
+
+    @AfterEach
+    void teardown() {
+
+    }
+
+    @Test
+    void test() {
+
+
+    }
+
+}
