--- parser/src/test/java/java/dj/djm/parser/ParserTest.java
@@ -1,30 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package java.dj.djm.parser;
-
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class ParserTest {
-
-    @BeforeEach
-    void setup() {
-
-    }
-
-    @AfterEach
-    void teardown() {
-
-    }
-
-    @Test
-    void test() {
-
-
-    }
-
-}
